Move getPeerDependenciesFromRegistry to /src/lib
