+++ src/versionmanager.ts
@@ -13,6 +13,7 @@ import isUpgradeable from './lib/isUpgradeable'
 import queryVersions from './lib/queryVersions'
 import getPackageManager from './lib/getPackageManager'
 import upgradeDependencies from './lib/upgradeDependencies'
+import getPeerDependenciesFromRegistry from './lib/getPeerDependenciesFromRegistry'
 import { FilterPattern, GetVersion, IgnoredUpgrade, Index, Maybe, Options, PackageManager, PackageFile, Version, VersionDeclaration } from './types'
 
 /**
@@ -122,34 +123,6 @@ export async function getOwnerPerDependency(fromVersion: Index<Version>, toVersi
       [dep]: ownerChanged,
     }
   }, {} as Promise<Index<boolean>>)
-}
-
-/**
- * Get the latest or greatest versions from the NPM repository based on the version target.
- *
- * @param packageMap   An object whose keys are package name and values are version
- * @param [options={}] Options.
- * @returns Promised {packageName: peer dependencies} collection
- */
-async function getPeerDependenciesFromRegistry(packageMap: Index<VersionDeclaration>, options: Options) {
-  const packageManager = getPackageManager(options.packageManager)
-  if (!packageManager.getPeerDependencies) return {}
-
-  const numItems = Object.keys(packageMap).length
-  let bar: ProgressBar
-  if (!options.json && options.loglevel !== 'silent' && options.loglevel !== 'verbose' && numItems > 0) {
-    bar = new ProgressBar('[:bar] :current/:total :percent', { total: numItems, width: 20 })
-    bar.render()
-  }
-
-  return Object.entries(packageMap).reduce(async (accumPromise, [pkg, version]) => {
-    const dep = await packageManager.getPeerDependencies!(pkg, version)
-    if (bar) {
-      bar.tick()
-    }
-    const accum = await accumPromise
-    return { ...accum, [pkg]: dep }
-  }, {})
 }
 
 /**
@@ -352,6 +325,5 @@ module.exports = {
   getOwnerPerDependency,
 
   // exposed for testing
-  getPeerDependenciesFromRegistry,
   getIgnoredUpgrades,
 }
