+++ test/versionmanager.test.js
@@ -303,31 +303,4 @@ describe('versionmanager', () => {
     })
   })
 
-  describe('getPeerDependenciesFromRegistry', function () {
-    it('single package', async () => {
-      const data = await vm.getPeerDependenciesFromRegistry({ 'ncu-test-peer': '1.0' }, {})
-      data.should.deep.equal({
-        'ncu-test-peer': {
-          'ncu-test-return-version': '1.x'
-        }
-      })
-    })
-    it('single package empty', async () => {
-      const data = await vm.getPeerDependenciesFromRegistry({ 'ncu-test-return-version': '1.0' }, {})
-      data.should.deep.equal({ 'ncu-test-return-version': {} })
-    })
-    it('multiple packages', async () => {
-      const data = await vm.getPeerDependenciesFromRegistry({
-        'ncu-test-return-version': '1.0.0',
-        'ncu-test-peer': '1.0.0',
-      }, {})
-      data.should.deep.equal({
-        'ncu-test-return-version': {},
-        'ncu-test-peer': {
-          'ncu-test-return-version': '1.x'
-        }
-      })
-    })
-  })
-
 })
